# service-java/src/main/java/org/choreo/demo/luxury/hotels/dto/UpdateReservationRequest.java
package org.choreo.demo.luxury.hotels.dto;

public class UpdateReservationRequest {
    String checkinDate;
    String checkoutDate;

    public String getCheckinDate() {
        return checkinDate;
    }

    public void setCheckinDate(String checkinDate) {
        this.checkinDate = checkinDate;
    }

    public String getCheckoutDate() {
        return checkoutDate;
    }

    public void setCheckoutDate(String checkoutDate) {
        this.checkoutDate = checkoutDate;
    }
}